# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Category {
  id: string;
  name: string;
  slug: string;
  description: string;
  image: string;
  productCount: number;
}

export interface Specification {
  label: string;
  value: string;
}

export interface Product {
  id: string;
  slug: string;
  name: string;
  description: string;
  category: string;
  categorySlug: string;
  images: string[];
  price: number;
  originalPrice: number;
  discountPercent: number;
  rating: number;
  reviewCount: number;
  stock: number;
  featured: boolean;
  bestSeller: boolean;
  deal: boolean;
  /** @nullable */
  badge: string | null;
  colors: string[];
  sizes: string[];
  specifications: Specification[];
  shipping: string;
  returnPolicy: string;
}

export interface ProductList {
  items: Product[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface HomePayload {
  featuredCategories: Category[];
  trendingProducts: Product[];
  bestSellers: Product[];
  deals: Product[];
  dealEndsAt: string;
}

export interface CartItem {
  product: Product;
  /** @minimum 1 */
  quantity: number;
  lineTotal: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  shipping: number;
  tax: number;
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  quantity: number;
}

export interface CartQuantityInput {
  /**
     * @minimum 0
     * @maximum 20
     */
  quantity: number;
}

export type CheckoutQuoteInputShippingMethod = typeof CheckoutQuoteInputShippingMethod[keyof typeof CheckoutQuoteInputShippingMethod];


export const CheckoutQuoteInputShippingMethod = {
  standard: 'standard',
  express: 'express',
} as const;

export interface CheckoutQuoteInput {
  shippingMethod: CheckoutQuoteInputShippingMethod;
  /** @nullable */
  coupon?: string | null;
}

export interface CheckoutQuote {
  subtotal: number;
  shipping: number;
  discount: number;
  tax: number;
  total: number;
  /** @nullable */
  couponApplied: string | null;
}

export type CheckoutInputShippingMethod = typeof CheckoutInputShippingMethod[keyof typeof CheckoutInputShippingMethod];


export const CheckoutInputShippingMethod = {
  standard: 'standard',
  express: 'express',
} as const;

export type CheckoutInputPaymentMethod = typeof CheckoutInputPaymentMethod[keyof typeof CheckoutInputPaymentMethod];


export const CheckoutInputPaymentMethod = {
  card: 'card',
  upi: 'upi',
  netbanking: 'netbanking',
  cod: 'cod',
} as const;

export interface CheckoutInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 3 */
  address: string;
  /** @nullable */
  apartment?: string | null;
  /** @minLength 2 */
  city: string;
  /** @minLength 2 */
  state: string;
  /** @minLength 2 */
  country: string;
  /** @minLength 3 */
  postalCode: string;
  shippingMethod: CheckoutInputShippingMethod;
  paymentMethod: CheckoutInputPaymentMethod;
  /** @nullable */
  coupon?: string | null;
}

export interface OrderItem {
  product: Product;
  quantity: number;
  lineTotal: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface Order {
  id: string;
  date: string;
  status: OrderStatus;
  shippingStatus: string;
  /** @nullable */
  trackingNumber: string | null;
  items: OrderItem[];
  total: number;
  customerName: string;
  customerEmail: string;
  paymentStatus: string;
  shippingAddress: string;
  estimatedDelivery: string;
}

export type AdminSummaryBestSellingProductsItem = {
  name: string;
  units: number;
  revenue: number;
};

export interface AdminSummary {
  totalSales: number;
  orders: number;
  customers: number;
  revenue: number;
  averageOrderValue: number;
  bestSellingProducts: AdminSummaryBestSellingProductsItem[];
  recentOrders: Order[];
}

/**
 * Not found
 */
export type NotFoundResponse = Error;

export type SearchParameter = string;

export type ListProductsParams = {
search?: SearchParameter;
category?: string;
featured?: boolean;
bestSeller?: boolean;
deal?: boolean;
/**
 * @minimum 0
 */
minPrice?: number;
/**
 * @minimum 0
 */
maxPrice?: number;
/**
 * @minimum 0
 * @maximum 5
 */
minRating?: number;
sort?: ListProductsSort;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 48
 */
pageSize?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  featured: 'featured',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating: 'rating',
  newest: 'newest',
  bestselling: 'bestselling',
} as const;