# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface CartItem {
  product: Product;
  /** @minimum 1 */
  quantity: number;
  lineTotal: number;
}