# lib/api-zod/src/generated/types/cartItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemInput {
  productId: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  quantity: number;
}