# lib/api-zod/src/generated/types/homePayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { Product } from './product';

export interface HomePayload {
  featuredCategories: Category[];
  trendingProducts: Product[];
  bestSellers: Product[];
  deals: Product[];
  dealEndsAt: Date;
}