# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  date: Date;
  status: OrderStatus;
  shippingStatus: string;
  /** @nullable */
  trackingNumber: string | null;
  items: OrderItem[];
  total: number;
  customerName: string;
  customerEmail: string;
  paymentStatus: string;
  shippingAddress: string;
  estimatedDelivery: string;
}