# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface OrderItem {
  product: Product;
  quantity: number;
  lineTotal: number;
}