# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Specification } from './specification';

export interface Product {
  id: string;
  slug: string;
  name: string;
  description: string;
  category: string;
  categorySlug: string;
  images: string[];
  price: number;
  originalPrice: number;
  discountPercent: number;
  rating: number;
  reviewCount: number;
  stock: number;
  featured: boolean;
  bestSeller: boolean;
  deal: boolean;
  /** @nullable */
  badge: string | null;
  colors: string[];
  sizes: string[];
  specifications: Specification[];
  shipping: string;
  returnPolicy: string;
}